tests/test_playwright.py
# Copyright 2026 James G Willmore
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

import re
from playwright.sync_api import Page, expect


def test_has_title(page: Page, base_url: str):
    page.goto(f"{base_url}/")

    # Expect a title "to contain" a substring.
    expect(page).to_have_title(re.compile("NiceGUI"))
