pyproject.toml
[project]
name = "pytasktracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arrow>=1.4.0",
    "click>=8.3.1",
    "dataclass-click>=1.0.4",
    "duckdb>=1.4.4",
    "duckdb-engine>=0.17.0",
    "fastapi>=0.129.2",
    "loguru>=0.7.3",
    "nicegui>=3.7.1",
    "pytest-order>=1.3.0",
    "sqlmodel>=0.0.34",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "bumpver>=2025.1131",
    "coverage>=7.13.4",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.7.2",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
# Set line length
line-length = 88
# Set minimum Python version
target-version = "py312"
# Exclude specific directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".venv",
    "__pycache__",
    "node_modules",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "I",    # isort
    "D",    # pydocstyle (docstring conventions)
]
# Ignore specific rules
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "D100",  # missing docstring in public module
    "D104",  # missing docstring in public package
]

# Allow fixing specific rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Use trailing commas
skip-magic-trailing-comma = false
# Auto-detect line ending
line-ending = "auto"
# Enable code formatting in docstrings
docstring-code-format = true

[tool.ty.environment]
python-version = "3.12"
python-platform = "linux"

[tool.ty.src]
include = ["src"]
respect-ignore-files = true

[tool.ty.rules]
# Uncomment to promote warnings to errors:
# all = "error"

[tool.ty.terminal]
error-on-warning = false

[tool.bumpver]
current_version = "0.1.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "v{new_version}"
tag_scope = "default"
"pyproject.toml" = ['current_version = "{version}"']
"src/mods/__init__.py" = ['__version__ = "{version}"']
